[Software] Added a fast display rendering mode to the Chip-8 interpreter.
Instead of transferring the frame buffer to the display at each DRW instruction call,
transfer it at a 60Hz rate and only when a new DRW instruction has been executed.

This allows to display a tremendous amount of sprites per seconds, but breaks the games
that rely on the DRW instruction to add some timing.

This new mode also relieves a bit the power consumption by soliciting less the display.